+++ test/unit_tests/test_ops.py
@@ -14,13 +14,8 @@
 import math
 
 import numpy as np
-import pennylane as qml
 import pytest
-from braket.circuits import gates, observables
-from braket.circuits.result_types import Expectation, Probability, Sample, Variance
-from pennylane.operation import ObservableReturnTypes
-from pennylane.tape.measure import MeasurementProcess
-from pennylane.wires import Wires
+from braket.circuits import gates
 
 from braket.pennylane_plugin import (
     ISWAP,
@@ -34,7 +29,6 @@ from braket.pennylane_plugin import (
     CPhaseShift01,
     CPhaseShift10,
 )
-from braket.pennylane_plugin.translation import translate_result_type
 
 testdata = [
     (CPhaseShift, gates.CPhaseShift, [math.pi]),
@@ -54,57 +48,3 @@ testdata = [
 def test_matrices(pl_op, braket_gate, params):
     """Tests that the matrices of the custom operations are correct."""
     assert np.allclose(pl_op._matrix(*params), braket_gate(*params).to_matrix())
-
-
-pl_return_types = [
-    ObservableReturnTypes.Expectation,
-    ObservableReturnTypes.Variance,
-    ObservableReturnTypes.Sample,
-]
-
-braket_results = [
-    Expectation(observables.H(), [0]),
-    Variance(observables.H(), [0]),
-    Sample(observables.H(), [0]),
-]
-
-
-@pytest.mark.parametrize("return_type, braket_result", zip(pl_return_types, braket_results))
-def test_translate_result_type_observable(return_type, braket_result):
-    """Tests if a PennyLane return type that involves an observable is successfully converted into a
-    Braket result using translate_result_type"""
-    obs = qml.Hadamard(0)
-    obs.return_type = return_type
-    braket_result_calculated = translate_result_type(obs)
-
-    assert braket_result == braket_result_calculated
-
-
-def test_translate_result_type_probs():
-    """Tests if a PennyLane probability return type is successfully converted into a Braket
-    result using translate_result_type"""
-    mp = MeasurementProcess(ObservableReturnTypes.Probability, wires=Wires([0]))
-    braket_result_calculated = translate_result_type(mp)
-
-    braket_result = Probability([0])
-
-    assert braket_result == braket_result_calculated
-
-
-def test_translate_result_type_unsupported_return():
-    """Tests if a NotImplementedError is raised by translate_result_type for an unknown
-    return_type"""
-    obs = qml.Hadamard(0)
-    obs.return_type = None
-
-    with pytest.raises(NotImplementedError, match="Unsupported return type"):
-        translate_result_type(obs)
-
-
-def test_translate_result_type_unsupported_obs():
-    """Tests if a TypeError is raised by translate_result_type for an unknown observable"""
-    obs = qml.S(wires=0)
-    obs.return_type = None
-
-    with pytest.raises(TypeError, match="Unsupported observable"):
-        translate_result_type(obs)
